mrcroqueto/apps/api/views/product_views.py
from rest_framework import viewsets
from apps.api.serializers.product_serializer import ProductSerializer


class ProductViewSet(viewsets.ModelViewSet):
    serializer_class = ProductSerializer

    def get_queryset(self):
        return self.get_serializer().Meta.model.objects.filter(state=True)
