mrcroqueto/requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
dj-static==0.0.6
Django==5.0.1
django-ckeditor==6.7.0
django-import-export==3.3.5
django-js-asset==2.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==21.2.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
psycopg==3.1.16
psycopg2==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
static3==0.7.0
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.4
xlrd==2.0.1
xlwt==1.3.0
